generic Skellam

+++ src/univariate/discrete/skellam.jl
@@ -35,7 +35,7 @@ end
 
 Skellam(μ1::T, μ2::T) where {T<:Real} = Skellam{T}(μ1, μ2)
 Skellam(μ1::Real, μ2::Real) = Skellam(promote(μ1, μ2)...)
-Skellam(μ1::Integer, μ2::Integer) = Skellam(Float64(μ1), Float64(μ2))
+Skellam(μ1::Integer, μ2::Integer) = Skellam(float(μ1), float(μ2))
 Skellam(μ::Real) = Skellam(μ, μ)
 Skellam() = Skellam(1.0, 1.0)
 
@@ -65,12 +65,12 @@ kurtosis(d::Skellam) = 1 / var(d)
 
 #### Evaluation
 
-function logpdf(d::Skellam, x::Int)
+function logpdf(d::Skellam, x::Integer)
     μ1, μ2 = params(d)
     - (μ1 + μ2) + (x/2) * log(μ1/μ2) + log(besseli(x, 2*sqrt(μ1)*sqrt(μ2)))
 end
 
-pdf(d::Skellam, x::Int) = exp(logpdf(d, x))
+pdf(d::Skellam, x::Integer) = exp(logpdf(d, x))
 
 function mgf(d::Skellam, t::Real)
     μ1, μ2 = params(d)
